+++ CHANGELOG.md
@@ -25,6 +25,8 @@
 - Enh #335: Replace `DbArrayHelper::getColumn()` with `array_column()` (@Tigrov)
 - Bug #338: Explicitly mark nullable parameters (@vjik)
 - Enh #337: Move `JsonExpressionBuilder` and JSON type tests to `yiisoft/db` package (@Tigrov)
+- New #319: Add parameters `$ifExists` and `$cascade` to `CommandInterface::dropTable()` and
+  `DDLQueryBuilderInterface::dropTable()` methods (@vjik)
 
 ## 1.2.0 March 21, 2024